backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-allauth==0.57.0
django-filter==24.1
drf-spectacular==0.27.1
python-dotenv==1.0.1
cryptography==44.0.2
Pillow==10.2.0
django-environ==0.11.2
psycopg2-binary==2.9.9  # For PostgreSQL support
gunicorn==21.2.0  # For production deployment 